reorganize editor; add rough spinner component
