+++ packages/ketcher-react/src/script/ui/views/components/StructEditor/StructEditor.jsx
@@ -16,7 +16,9 @@
 
 import { upperFirst } from 'lodash/fp'
 import React, { Component, createRef } from 'react'
-import Editor from '../../editor'
+import Editor from '../../../../editor'
+
+import styles from './StructEditor.module.less'
 
 //TODO: need to update component after making refactoring of store
 function setupEditor(editor, props, oldProps = {}) {
@@ -59,8 +61,8 @@ class StructEditor extends Component {
     this.logRef = createRef()
   }
 
-  shouldComponentUpdate() {
-    return false
+  shouldComponentUpdate(nextProps) {
+    return this.props.indigoVerification !== nextProps.indigoVerification
   }
 
   componentWillReceiveProps(props) {
@@ -109,15 +111,22 @@ class StructEditor extends Component {
       onAromatizeStruct,
       onDearomatizeStruct,
       onAttachEdit,
+      indigoVerification,
       ...props
     } = this.props
+
     return (
       <Tag
-        onMouseDown={ev => ev.preventDefault()}
+        onMouseDown={event => event.preventDefault()}
         {...props}
         ref={this.editorRef}>
         {/* svg here */}
         <div className="measure-log" ref={this.logRef} />
+        {indigoVerification && (
+          <div className={styles.spinner_overlay}>
+            <p>spinner</p>
+          </div>
+        )}
       </Tag>
     )
   }
