+++ packages/ketcher-react/src/script/ui/views/components/StructEditor/StructEditor.module.less
@@ -1,0 +1,9 @@
+.spinner_overlay {
+  position: absolute;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
